+++ index.js
@@ -1,4 +1,5 @@
 const server = require("./api/server.js");
+require('dotenv').config();
 
 const PORT = process.env.PORT || 5000;
 server.listen(PORT, () => console.log(`\n** Running on port: ${PORT} **\n`));
