+++ knexfile.js
@@ -4,10 +4,14 @@ module.exports = {
 
   development: {
     client: 'sqlite3',
+        useNullAsDefault: true,
     connection: {
       filename: './data/users.db3'
     },
-    useNullAsDefault: true,
+    pool: {
+      afterCreate: (conn, done) => {
+        conn.run('PRAGMA foreign_keys = ON', done);
+      },
     migrations: {
       directory: './data/migrations'
     },
