+++ package.json
@@ -20,9 +20,13 @@
   },
   "homepage": "https://github.com/lizdoyle/webauth-i-challenge#readme",
   "dependencies": {
+    "bcrypt": "^3.0.7",
     "bcryptjs": "^2.4.3",
+    "connect-session-knex": "^1.5.0",
     "cors": "^2.8.5",
+    "dotenv": "^8.2.0",
     "express": "^4.17.1",
+    "express-session": "^1.17.0",
     "helmet": "^3.21.2",
     "knex": "^0.20.4",
     "nodemon": "^2.0.1",
